Rename parser.py to parsers.py
