+++ tests/test_integration_parse_player_advanced_season_totals.py
@@ -6,7 +6,7 @@ from lxml import html
 from basketball_reference_web_scraper.data import Team, Position, TEAM_ABBREVIATIONS_TO_TEAM, \
     POSITION_ABBREVIATIONS_TO_POSITION
 from basketball_reference_web_scraper.html import PlayerAdvancedSeasonTotalsTable
-from basketball_reference_web_scraper.parser import PlayerAdvancedSeasonTotalsParser, PositionAbbreviationParser, \
+from basketball_reference_web_scraper.parsers import PlayerAdvancedSeasonTotalsParser, PositionAbbreviationParser, \
     TeamAbbreviationParser
 
 season_2019_totals_html = os.path.join(os.path.dirname(__file__), './NBA_2019_advanced_totals.html')
