+++ tests/test_integration_parse_player_box_scores.py
@@ -7,7 +7,7 @@ from basketball_reference_web_scraper.data import TEAM_ABBREVIATIONS_TO_TEAM, LO
     OUTCOME_ABBREVIATIONS_TO_OUTCOME
 from basketball_reference_web_scraper.data import Team, Outcome
 from basketball_reference_web_scraper.html import DailyLeadersPage
-from basketball_reference_web_scraper.parser import TeamAbbreviationParser, \
+from basketball_reference_web_scraper.parsers import TeamAbbreviationParser, \
     LocationAbbreviationParser, OutcomeAbbreviationParser, \
     SecondsPlayedParser, PlayerBoxScoresParser
