+++ tests/test_integration_parse_player_season_totals.py
@@ -6,7 +6,7 @@ from lxml import html
 from basketball_reference_web_scraper.data import Team, Position, POSITION_ABBREVIATIONS_TO_POSITION, \
     TEAM_ABBREVIATIONS_TO_TEAM
 from basketball_reference_web_scraper.html import PlayerSeasonTotalTable
-from basketball_reference_web_scraper.parser import PositionAbbreviationParser, TeamAbbreviationParser, \
+from basketball_reference_web_scraper.parsers import PositionAbbreviationParser, TeamAbbreviationParser, \
     PlayerSeasonTotalsParser
 
 season_2001_totals_html = os.path.join(os.path.dirname(__file__), './NBA_2001_totals.html')
